Restructure baclend code
