+++ quarkus-backend/src/main/java/models/RenameForm.java
@@ -1,13 +1,10 @@
-package org.ugomes.controllers.rest;
+package org.ugomes.models;
 
 import java.io.InputStream;
 
 import javax.ws.rs.FormParam;
 import javax.ws.rs.core.MediaType;
 import java.io.File;  // Import the File class
-import org.jboss.resteasy.annotations.providers.multipart.PartType;
-import org.jboss.resteasy.annotations.providers.multipart.MultipartForm;
-import org.jboss.resteasy.plugins.providers.multipart.MultipartFormDataInput;
 
 public class RenameForm {
     @FormParam("fileDir")
